pyproject.toml
[project]
name = "udf-halloween"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Henrique Kirch Heck", email = "me@henriquekh.dev.br" }]
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.4",
    "pygame-ce",
]

[project.scripts]
bullet-cemitary = "bullet_cemitary:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["bullet_cemitary"]

[tool.uv.sources]
pygame-ce = { git = "ssh://git@github.com/pygame-community/pygame-ce.git", tag = "2.5.6" }

[tool.ruff]
select = ["ALL"]
ignore = ["D", "T201", "TC002", "ERA001", "C901"]
